Move files to folders and remove useSession
